requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
qrcode==7.3.1
requests==2.27.1
soupsieve==2.3.2
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
